spi_qsd: Fix for Input overrun and Invalid fifo size errors
The Input overrun errors are due to a bug in the controller.
This bogus error is reported when a transition from run
to reset state occurs and if the input FIFO has an odd number
of entries. We simply ignore such errors when the
transfer has completed successfully.

Signed-off-by: Harini Jayaraman <harinij@codeaurora.org>
(cherry picked from commit 29e7f7a8d507fc2386703cbfdb472e4a2fd0b274)

Change-Id: I9bb1d2ee7f53a2fe97ac9e9c350e7f423f64f8bb
CRs-fixed: 273258
Signed-off-by: Harini Jayaraman <harinij@codeaurora.org>