USB: msm_otg: Fix the bug in PHY suspend failure recovery
When PHY suspend fails, OTG state machine is initialized to UNDEFINED
state and work is scheduled.  State machine work takes care of resetting
the PHY and initializing the inputs.  Peripheral or host controller
driver is activated based on the B-session valid and Id inputs.  But this
recovery mechanism with wall charger stops charging and activates the
peripheral controller driver which was already in active state.  Handle
the wall charger scenario in UNDEFINED state by changing the state to
B_PERIPHERAL and schedule the work to take care of resetting the PHY and
putting the hardware in low power mode.

CRs-Fixed: 275374
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
(cherry picked from commit 2972bb7b89590c4d50801fa04e4b6024cbde0fe1)

Change-Id: Ie6674972030083f310e2fde30566c7038e2df02a
Signed-off-by: Sunil Joseph <sunilj@codeaurora.org>