msm: kgsl: sync outer cache when submitting ringbuffer commands
On msm 7x27, the outer cache must be synced as part of the mb()
macro. Otherwise the ringbuffer can be in an inconsistent state
in RAM when the write pointer is updated.

This is an interim change until we can safely disable
CONFIG_ARCH_HAS_BARRIERS and use the arm version mb() which calls
outer_sync() instead of the msm version which does not.

(cherry picked from commit b6bee25c3e143abd5875ea8398a62c39d2745d64)

Change-Id: Ie8909f9dae1e544e9846a40a588afc32374d3db2
CRs-fixed: 273669
CRs-fixed: 274294
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>