[7x27_common][USB] Send uevent whenever a function is enable/disable.
In BSP original, If a function is enable(or disable) when USB cable
is connected, Android gadget does not send uevent to user space.
As a result, Android platform cannot oberve the uevent. By this commit,
when a function is enable/disable, send uevent for enabling/disabling
function and Andoird can receive the uevent.

Change-Id: Iec46329035adf5af1abb07abad031a767422ea5e