power: pmic8058-charger: force VCHG to 0 volts only once
While going to suspend, when the usb cable is removed, multiple
invocations of chgval interrupt were seen. This chgval interrupt is a
wakeup interrupt which causes the system to wakeup from suspend.

In the chgval interrupt we write to test register to force VCHG
to 0 volts. This forcing further causes activation of chgval interrupt.
So force the VCHG to 0 volts only the first time we detect that the
cable is removed. Ignore the rest of the invocations.

(cherry picked from commit 4c0b8bfd04c86d21b7dbbe0507d3117d37509574)

Change-Id: Ie875981968c47d6c10659b19db60d0c870cf38f2
CRs-Fixed: 274343
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>