GIC: implement disable callback
The disable callback is required to be implemented so that we indicate
to the mpm that a particular interrupt has been disabled. Doing this
helps the mpm stay current with interrupt status allowing it to choose
the right low power mode while going in to idle.

Note that we do not mask the interrupt in hardware, taking advantage of
lazy disabling and IRQ_PENDING checks in check_wakeup_irqs() while
going to suspend.

(cherry picked from commit 6963a62e12836fe282bac41e7a7a40b1e0533890)

Change-Id: I23b4fb2987cd4845aed6db594f2036e7a64d1278
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>