msm: gpio-v2: implement disable callback
The disable callback is required to be implemented so that we indicate
to the mpm that a particular interrupt has been disabled. Doing this
helps the mpm stay current with interrupt status allowing it to choose
the right low power mode while going in to idle.

Note that we do not mask the interrupt in hardware, taking advantage of
lazy disabling and IRQ_PENDING checks in check_wakeup_irqs() while
going to suspend.

(cherry picked from commit 448d755a1e45668f82bc615e023a730103ef94aa)

Change-Id: If333698895122b5f16a51e48ee81363cce8e3d0e
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>