ASoC: msm8x60: Fix the afe_loopback triggering issue
Two variables used to cache the dev_ids are not initialized in init
function. These two variables are taking zero as default value, which is
still valid device id for handset_rx. Due to this when we enable/disable
the handset_rx, it is triggering AFE_Enable. Fix this issue by initializing
these variables to DEVICE_IGNORE in init function.

Change-Id: Iec065c62bd4e7bfb2470068d82d39252734a4306
Signed-off-by: Deepa Madiregama <dmadireg@codeaurora.org>