msm: watchdog: Add parameter to configure who handles bark interrupts
Add an "appsbark" flag that determines who handles watchdog barks that
come in. If it is set to 0(the default), the secure side will handle
barks, and save the registers out to a buffer in the ramdump. If it is
set to 1, Linux will handle barks, and dump the stack traces of all
processes upon receiving a bark.

Additionally, the secure side requires that the interrupt be configured
before sending the scm call.

(cherry picked from commit 6334be5d62a34414f14c58c46cb923c3b3d3b45b)

Change-Id: I5e0637820d56dd2b333b74665676d4a04ca313be
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>