msm: clock-7x30: Select LPXO source when disabling clocks with pwr rails
For clocks that have modem-controlled power rails associated with them,
a clocked input must remain selected even when the clock is disabled.

The power rail APIs are meant to be called while the clock is disabled.
Because the modem may assume that a valid clock source is selected when
it performs the power rail control operations, the clock driver should
not select a grounded input. Instead, when disabling the clock, LPXO
will be selected for good power savings.

Change-Id: I9d0923268948b9d3b67538a0d946570ca025cfc8
CRs-Fixed: 280184
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>