msm: kgsl: don't call kgsl_pwrctrl_wake() from isr context
This function makes sure that the power comes on before we
try to touch the hardware via kgsl_regread and kgsl_regwrite,
which is usually a good thing.  Except that there can races
with with the isr which can cause us to call blocking functions
like requesting bus bandwith. This change makes a separate
set of regread/regwrite functions that don't check that the
clocks are on. These should be used from the isr, because
clocks should already be on if we're getting valid interrupts.

This change also unravels the 2d regread/regwrite functions
which were too convoluted.

Change-Id: I0cb0ea3b342f78fad1ad06d929f13a630dcb7cfa
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>