camera: 8k: removed gpio reset request of front camera sensor gbrel.
The front camera sensor ov7692 does not have the reset pin connected,
the standby/power pin will do the reset. proper release of reset
for the back sensor is added. The standby pin and the reset pin for
the back sensor is pulled down rather than in the floating condition.

Change-Id: Icbf3eb32c221a7638b56340eba7f22b878b9ac70
Signed-off-by: Jeyaprakash Soundrapandian <jsound@codeaurora.org>