msm: dma: support dma power management on msm8x60
Move the initialization currently done in the dma driver's init function
to a probe function, and make the datamover a proper platform_device on
all targets. This is done so that each adm has its own struct device
that it can pass to the clock api. Additionally, move the irq and base
address fields of the dmov_conf structure to platform data, so that
there is no need to rely on consistent irq number and base address
macro naming.

This change makes it so we actually vote for the ADM clocks on msm8x60,
rather than relying on them coming up on.

(cherry picked from commit e583ac0df1be81e1454cf0190317835446d8386a)

Conflicts:

	arch/arm/mach-msm/dma.c

Change-Id: Icf3a1be3e4e3f9bd98767a2bd03b96493863546d
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>